Add guest scaffolding (controller, routes and views)
Part of the work to build out @stevepetebruce's designs into the app,
this is the first step in having a guest's show view, which is accessible
when clicking 'view more' in the trip show view.

Subsequent work will add the actual html and styles of the design,
overwriting this scaffolding boilerplate.