Add created_by_id and updated_by_id to users and trips_users.
We will want to record which user created or updated:
A user (both guides and guests);
A trip (already has these fields);
And a trips_user (the join between a trip and a user).

This is to cover the situation where there is a query regarding
who changed what data on a trip, user, or who added a user to a trip, etc.

Subsequent work will investigate using Rails' CurrentAttributes or just
Devise style current_user, and also add Devive authentication library.
Then we will have the concept of a current user in the app, which we can
use to update these fields.

Refs:
https://github.com/Book-Your-Place/bookyourplace/issues/19
https://github.com/Book-Your-Place/bookyourplace/issues/20