Replace Users with Guests and Guides.
It's possible that a Guide may be a Guest (on another trip).
Or a Guest of one trip may become a Guide on another trip.
Having a single table of Users which then extend into Guests
and Guides would not of allowed this. So needed to break it out
into two distinct tables / models not extended from User.