Add created_by and updated_by fields to guests and guides.
We want to know who has created a guest or a guide (less likely it will
be anyone other than the guide themselves but just for consistency).

A guest could be created by themselves (when they sign up for a trip),
or by a guide who has added them to a trip.

Subseuqnt work will add a controller concern to seamlessly update
this two fields when the model is created or updated.