Add Stripe gem.
This is the first step in allowing the app to accept payments in
the process of creating a booking.

Subsequent work will expose the Stripe library, build out forms for
users' payment details, addition of .env to manage sensitive config
variables (in this case the Stripe test API key).

Finally we will need to find a way to work with multiple accounts via
Stripe's Connect API. We will need to set up accounts for lots of
different guides and have access to them to make payments.