Add organisation model.
Part of the work to incorporate Stripe API.

We are only building out the specific models from the entity
relationship diagram as and when required in the app layer.

When a booking is created we need to know which connected account
to pay the charge into.
An organisation has a stripe_account_id which we will use to make
this payment.

Subsequent work will add the models and their associations with
this organisation model.