Use bookings as join table between guests and trips.
Part of the work to allow a booking to create a guest and
associate that guest with a trip.

It became clear that the join_table 'guests_trips' was a duplication
of the 'booking' table model. The booking table will have / has more
required functionality than a simple join table, so makes sense to
delete the guest_trips table and fix up all macros and specs to reflect
this.