Allow guests to be created without a password.
Part of the work to allow a booking to create a guest and
associate that guest with a trip.

Devise, by default, forces all its models to have their password field set.
This doesn't make sense in this case, when a guest is automatically created
by the scenes, when a booking is created.

Subsequent work will send an email out to a guest, to make sure they
have confirmed their email and set a password, before they can perform
any more actions, such as book another trip, any kind of payment, edit
guest, etc.