Find or create guest when creating a new booking.
Part of the work to allow a booking to create a guest and
associate that guest with a trip.

Once a user has used the public booking page to make a payment,
we get the stripe token from Stripe and post to bookings#create.
This then uses the email in the bookings#new form to find or create
a user and associate them with the booking and thus the trip.

Subsequent work will use the stripe token to actually make the charge
on the guest's card and funnel that over to the trip owner's
organisation connected/destination account.