Add public/guests controller and views.
Part of the work to allow a booking to create a guest and
associate that guest with a trip.

Once a booking has been made, we want to get the new guest to
add their extra details (address, phone number, etc) to their
guest record so the guide can access this later.

For security reasons, the new guest can only access this public
guest/edit view for the first 30 minutes after the guest is created.

Subsequent work will add a public/trips/:trip_id page so that, if
the trip_id is present in the params, we can redirect the guest back to
that page, or perhaps the guest log in page. Still to be decided.