Create organisation_membership join table.
Part of the work to allow guides to share a public link to make
a new booking with a member of public.

We need to associate organisations with guides so that we can use
the stripe_account_id associated with an organisation to make a
charge to that organisation's stripe destinaition account.

Subsequent work will associate the trip -> owner (guide) ->
organisation, so we can access the relevant organisation associated
with the trip's new booking, when the booking is created.