Add organisation to trips.
Previous work created the OrganisationMemberships join table
between organisations and guides, which had a boolean owner field
to indicate if that guide was the owner of that organisation.

But when creating a new booking, where we only have the reference
to the trip, there is no way to find out which guide (associated with
the trip) was the owner of the correct organisation (the organisation
that owned this trip.)

This work fixes that by adding a simple refernce to the organisation
to the trip model.

Subsequent work, may remove the owner field from the
OrganisationMembership model and just add a simpler owner/ guide
reference directly on the organisation model.