Add SideKiq gem for background job processing.
Part of the work to charge guests and make payments to guides.
We want to remove all calls to external services (ie: Stripe)
out of the web request cycle and into background jobs.

Subsequent work will include creating a ActiveJob for make requests
to Stripe and UX regarding letting the guest know if the charge has
been successful or not.