Create Plan model.
Part of the work to allow guests to make a payment to a guide.

A guide's organisation will be associated with a plan (via a
subscription), which will be used to apply an application/platform
fee to each charge, which will paid to our account.

Subsequent work will add subscription model, the join between plans
and organisations.