Add subscription model.
Part of the work to allow guests to make a payment to a guide.

A guide's organisation will have a subscription to a plan. We will
then use the details of the plan (flat fee amount, etc) to make an
application/ plaftorm charge to our account when a guest pays for
a full booking.

It may be useful to keep track of all the subscriptions/ plans a
guide has had, so we will allow multiple subscriptions to exist as
a record of plan changes, but there will only ever be one
current_subscription / plan for an organisation, which we will use
as the basis of the platform charge.

Subsequent work will expose the subscription / current_subscription
and plan in the organisation model.