Add subscription to organisation and plan models in the app layer.
Part of the work to allow guests to make a payment to guides.

This work continues on from the work where subscriptions were added.
This exposes subscriptions to organisations and plans in the app layer
and associated test coverage.

Subsequent work will access the trip's organisation's plan to make
a platform charge to our account when we handle the payment from a
guest to a guide.