Add dot env in the front end.
We were already using the dot env gem to manage env vars in the backend
but we need to access Stripe keys in the front end too. For that we needed
to add the dot env yarn package.

Both the backend and front end dot env use the .env file env vars in the
same way.

Ref: https://github.com/motdotla/dotenv