Add Stripe API keys.
Part of the work to allow guests to make a new bookings and a payment
to a guide. We need to access our Stripe API keys (both publishable -
for use in the front end with Stripe.js and secret - for use on the
server).

We don't want to commit these to github, so need to add them to the
.env file.

Follows on from previous work that added back and front end dot env
gem/ yarn package.