Surface the new email field on booking model in the app layer. Delegate organisation_name to trip.
Part of the work to allow a publicly accessible booking URL.

Needed to separate the data accessible by unauthenticated users
from authenticated guests.

Duplicating the fields that should be on the guest model onto
the booking model, allows for a clear separation of these two
areas of data. Preventing anyone from being able to edit sensitive
guest data.

The organisation_name (delegated to trip model) is used in the
public booking views.

Also, some coding standards in the touched files.