Add Bookings::Payment and External::StripeApi service models.
Part of the work to allow a publicly accessible booking URL and
Stripe API integration.

These service models handle the creation of a charge via a new
booking.

The design is meant to allow loose-coupling between bookings and
Stripe API. In case we ever want to swap out the payment provider.

Bookings::Payment accepts a booking object and a token representing
the new customer. Which then uses the External::StripeApi to
communicate with the Stripe API.

Follow up work will include unit testing these models. They are
currently covered in the higer level request specs.