Refactor all public resources to only access bookings (rather than guests).
Part of the work to allow a publicly accessible booking URL.

Unauthenticated users will be able to access a public/trips/:trip_id/bookings/new
page, enter their card details and create a new booking on a trip.

This commit contains the final routes and controller (and associated specs)
changes that implement this.