Expose real data from trip and booking in the public/bookings/new view.
This is the last final part of the work to use real data, rather than
hard-coded data in the views.

The work is spread over a few models, tests & test factories and a single view
but the main change is in the view.

Some TODOs made where there's technical debt to pay down later.