Add email functionality to the app.
We want to sent an email to the email address assciated with a new booking
to confirm their booking and also (later) log in as a guest (not just a
public booking).

Follow up work will include:
Sending a "new booking" email to the guide
Using the guide's email address and name in the from and reply-to fields
Setting up background jobs using redis and sidekiq to actually send the
  ...emails in a background job.
Designing the email content using relevant content/data.
Using the status of the new booking's data completion in the email.