Tighten up GuestController actions and add sign_in / sign_out spec helpers.
This work reduces the actions in the GuestController to those we have to have
to increase security and prevent any destructive behaviour.

Part of the work separating the guest and guide routes.

Calls the Devise authentication method authenticate_guest! on all guest
routes, so a guest has to log in to visit their data.

Follow up work will add a one-time-only token-based log in URL, which
will be emailed to the email address in every new booking. Which, when
clicked creates a new guest (from the booking data).