Define the routes that guests and guides are directed to after they log in/ log out.
Following on from the work that separated out all guests and guide routes, this
work fixes the routing when guides or guests log in or out.

Previously, they would always be redirected to the generic log in page, when logging
in or out, which would be the guest log in page, due to its primary order in the
routes file, but now guests and guides are kept within their respective namespaces
and are directed to trips#index after logging in and guests/ guides /sign_in when
logged out.

Refs:
https://github.com/Book-Your-Place/bookyourplace/pull/93
https://github.com/Book-Your-Place/bookyourplace/pull/93/commits/a68fe00e8b762290115da786a1e4ae8bc2ced387