Add rb-readline gem to fix broken development environment.
During the work to set up SSL certificate for the app, it was nec
to install certbot (as per LetsEncrypt's instructions) to generate
the SSL certificates.

However certbot, for some reason, then breaks the development
environment, with an error message of, missing image: readline.

Installing this gem fixes that issue.

As it's only loaded in the development environment, it will not affect
production at all and as it's likely we will need to use certbot again
may aswell fix this now so that there's less confusion in future.