Fix intermittent failing spec caused by invalid organisation subdomain.
A number of specs were failing intermittently that were relying on the
organisation factory, which was creating an invalid subdomain for the
organisation, due to it being too short (< 3 chars).

This was exacerbated by the regex of the subdomain validation that
actually excluded subdomains that were 3 chars long. This has also been
fixed in this work.

Subsequent work will fix the intermittently failing specs in other areas.