New public booking form UX improvements - use autofocus on first form field.
Working on copying booking data to a guest (associated via email address)
it was noticed that we have first_name and last_name fields on the booking
model but have combined these fields into a single name field in the new
booking form.

Wanting to make sure that we weren't making any mistakes by this change in
design I followed these recommendations (1) which recommend, among other
things, that we autofocus the first field that the user needs to complete.

This change does that.

Subsequent work will change the booking fields (first_name and last_name) to
name.

(1) https://conversionxl.com/blog/form-design-best-practices/