Replace first_name and last_name with name field on Booking model.
Part of the work to copy over data from a new booking to a guest.
We have changed the design of the new bookings form (combining the
first_name and last_name into name) but not the backing model.

This also makes the booking model consitent with the guest model
(which has a single field for name).

Subsequent work will extend the guest model fields to allow for
name_booking, name_override and the final name field, so we can
track where exactly the guest's name has come from and prioritise
when the guest has set the value, over the booking's value.