Replace bookings' first_name and last_name fields with name field in app layer.
Part of the work to Part of the work to copy over data from a new booking to a guest.

We have changed the design of the new bookings form (combining the
first_name and last_name into name) but not the backing model.

This follows the work that changed this at the DB layer.

Subsequent work will extract the first name, when we need it, by splitting the
name field if there's two names in there and using only the first one.