Replace bookings' first_name and last_name fields with name field in test layer.
Part of the work to copy over data from a new booking to a guest.

We have changed the design of the new bookings form (combining the
first_name and last_name into name) but not the backing model.

This looks the only area where the tests failed after the chanage.