Add required fields to bookings and guests models/tables.
Part of the work to copy over data from a booking to an associated
guest, when the guest clicks on the email confirmation link.

There was a number of fields missing from the booking table and the
guest table that were required.

The pattern used being: the guest will have a field, 'x', with a corresponding
'x' field on the booking. Any data copied from the booking's x field will be
copied over the guest's x_booking field.

If there is no value for the x_override field (on the guest) then the guest's
x field will be set as the value for the x_booking. If there is a value for the
guest's x_override field, then that value will be set as the guest's x value.

This allows the priortitisation of the value set manually, by the guest, to
override the value from the booking for that, and all other fields that have
this pattern.

Subsequent work will implement this copying of data over from booking to
guest.