Expose, as enums, the allergies and dietary_requirements fields in the app layer.
Part of the work to copy over data from a booking to an associated
guest, when the guest clicks on the email confirmation link.

Previous work added lots of new fields required to record details of a new booking
and guest.

This just exposes some of the intention of the fields that are not freeform string/text
fields.

It is better to limit user choice when appropriate, so allergies and dietary_requirements
are a finite set of possibles, they can be limited to a range of values.

We can add to these options later as more cases come to light.

Subsequent work will allow the multiple selection of these values and change the allergies
free text field to a multi select UI.