Do not redirect_back to edit_booking view on time out of editing booking.
Whilst working in this area I noticed that you are stuck in an infinite loop
when you submit to update the booking that is out of time for editing.

By redirecting to the edit page, when you are out of date, the check_timeout
before_action filter kicks in again and redirects you back to the edit view,
infinitely.