Connect new booking fields with booking controller and booking edit form.
Part of the work to copy over data from a booking to an associated
guest, when the guest clicks on the email confirmation link.

This work allows the new fields in the controller params and tidies up
the booking edit form so only the nec. fields required for the booking
controller and body are present and all connected up.

Subsequent work will connect up the allergies and dietary_requirements
form fields with the booking's model enums, using checkboxes etc.