Add Guests::Updater to handle updating a Guest from a Booking model.
This work will allow any blank fields on a Guest to be updated from their
associated booking.

This will be run when the guest clicks on the confirmation link in the email
we send them.

Only then can the guest fields be updated to prevent anyone else making a booking
using their email address maliciously and corrupting their data.

Follow up work will plumb this in to be called when appropriate.