Rename Guests::Updater -> Guests::BookingUpdater.
Part of the work to copy over data from a booking to an associated
guest, when the guest clicks on the email confirmation link.

The logic for updating a guest from its _booking and _override fields
became increasingly complex so it made sense to extract it out of the
Guest model and into its own service model Guests::Updater.

The original class is now renamed Guests::BookingUpdater to allow for
this.