Include link to guest show view with one time login URL.
Part of the work to include a link in new booking email to guests.

This allows guests to log straight in to see details of the trip they
just booked on and their personal details.

Final work will make sure that when they have clicked this link, any
fields on the guest model that don't have data have their relevant booking
fields data copied over.