Update guest's booking fields based on most recent associated booking.
Part of the work to copy over a booking to associated guest data.

We only need to copy over the most recent booking's data, so can
simplify the Guests::BookingUpdater class and express that more
clearly.

Follow up work will fix the pending test, which is breaking because of
the update of enum fields on the guest (ex: dietary_requirements_booking))
which are being updated as the enum label (eg: 'vegan'), not the integer
representation.

Also, will call the updater in the guest's controller when a guest logs in
using the one time login link.