Copy over most recent booking's data when guest confirms email.
We want to copy over any data from the most recent booking to the
associated blank guest _booking fields when a guest confirms their email
by clicking on the one time login link in the email we send with each new booking.

Follow up work will fix the pending spec by changing how we deal with the guest's
fields that are enums. There's an issue with the bulk update process that needs fixing.

Need to convert the columns to strings from integers:
Ref: https://sipsandbits.com/2018/04/30/using-database-native-enums-with-rails/