Remove eunum validations to fix the copying over booking data to guest fields.
Enums in Rails behave strangely when you have validations applied to them.
When trying to update an enum field with it's mapped value, eg 'dairy', rather
than 0, you encounter validation errors, such as: 0 not included in list.
Even though that is the correct underlying value.

Futher work will remove these fields as enums and either include them as bitwise
flags or as a has_many association, to allow more than one option to be selected
and get away from this odd and restrictive behaviour of enums. They were not the
correct choice to represent this data.