Force SSL. redirect http -> https in production.
Part of the work to provide a link to get guides to log in to the app
via the marketing page (from: https://mybookingsherpa.com/, to login
into: https://app.mybookingsherpa.com/)

Currently http://app.mybookingsherpa.com/ is still accessible but
shows an error in the browser when you access it.

This should redirect http requests to https due to this change on
heroku.