Add Payment model.
Part of the work to use real data throughout the current views.
Which is part of the work to prepare for first user testing.

We need to show a guest's booking's status for a trip.

That can be one of many states, depending on:
Payment state: nothing paid, deposit paid or full amount paid.
Personal details state: minor details entered, or full details entered.

For these states to be updated we need to record the payment of deposits
and full payments per booking.

Follow up work will create a payment record when we've succesfully made
a charge through Stripe and add the Bookings::StatusUpdater service model.