Add Payments::Factory to handle the creation of payment records.
Part of the work to use real data in views in preparation for the
first user testing.

We will need to know the status of a booking. This will be based on:

1) The payment status (unpaid, deposit_paid or full_amount_paid)
2) Personal details status: basic personal details, or full details entered.

To know about the payment status we need to record any payments made with
a booking. This work allows that.

Follow up work will include creating a payment record in the bookings#create
controller action and then updating the booking's status based on this.