Add Bookings::StatusUpdater service model.
Part of the work to use real data in views in preparation for the
first user testing.

We have been using hard coded status colours in the trip index view.
This is work that will replace that with real data.

The booking's status has changed from pending / complete to three
possible statuses:
red, yellow and green.

The default is red. The status of a booking will change once
payments are made and guests update their personal details.

Once a full payment is made (not just a deposit) and all required
guest personal details are complete the status changes to green.

The guide can quickly see what needs chasing up based on these statuses.

Follow up work will call this class when a new payment is made or when a
booking/ guest is updated.