Use only yellow and green for booking status states.
Part of the work to use real data in views in preparation for the
first user testing.

A refactor of previous work, after discussion it was decided that only
two states (yellow and green) were necessary.

Green is everything required from the guest is complete (payment and
personal details) and yellow is action required/ something missing from
the booking's payment or personal details.

Follow up work will update a booking status when there's a new payment
or a booking has been updated.