Refactor Bookings::StatusUpdater.
Part of the work to replace hard data with real data
in views in preparation for the first user testing.

As the booking will be setting its own status field in a before_save callback,
we need to expose the new_status value, rather than just update the booking itself.
This might of caused an infinite loop.

As we also need to simply update the booking from the bookings controller, we still need
to have the #update method to allow for that use case.

re: #full_amount_paid?
Do we want to this to be true when:
total_paid == @booking.full_cost ... this would highlight when there's been an over payment... BUT...
OR: total_paid >= @booking.full_cost ...this would cover when there's extras to be paid for (equipment, etc)

Have set this to be total_paid >= @booking.full_cost for now.
This is due to the addition of extra costs (to be included later) which would start marking these
guest (who have paid for extras) falsely as incomplete, instead of complete when they have paid
more than the full price of the trip.

We can highlight in the UI the payments made so that guests and guides can see if there's been an
overpayment made.

Follow up work will call this class when a new booking / payment is made and when a booking is updated
(ie: with personal details).