Dynamically create fields on GuestDecorator that expose most recent booking's data if guest data is blank.
Part of the work to replace hard data with real data
in views in preparation for the first user testing.

If a guest has added data to their booking (address, name, etc) but not confirmed their email
address, then the guide will just be faced with lots of blank data as the copying of booking
data to guest fields only happens when the guest confirms their email.

While we are waiting for that confirmation step to occur we can at least use the most
recent booking's data so the guide has something to work with whilst still keeping the
guest data secure by requiring the confirmation of the email before any copying of data
can occur.