Rename Bookings::StatusUpdater -> Bookings::Status.
Part of the work to expose real data in the UI in prep for first
user testing.

The responsibilities of the Bookings::StatusUpdater have grown so
that it is no longer just updating the status of a booking but also
providing the state of a booking, ie: dietary_requirements?,
medical_condition?, etc.

This info will be used in the UI to alert the guide that a guest
has dietary_requirements?, medical_condition?, etc.