Remove references to guest.some_field when using BookingDecorator.
By using (for ex:) booking_decorator.guest.name, rather than
booking_decorator.name, we were referencing the guest.name field
directly on the guest, rather than using the dynamically created
fields on the booking_decorator, that gives precedence to the bookings'
guest field, then falls back to the booking's equivalent field.

This meant that even if a booking had a name value, it was showing up
blank as we were only exposing the guest field of name, not the overriden
name (booking.guest.name || booking.name) method on the booking_decorator.