Fix allergies and dietary_requirements enums on booking and guest models.
Discovered an issue when creating a new booking that prevented any
bookings with allergies and dietary_requirements set to none were failing.

As none is a reserved word, needed to rename. Adding the prefix: true
to all enums adds the name of the enum to the helper methods, ex:
allergies_none rather than none.

To make sure the none value behaves like the enum genuinely doesn't have
a value, it was necessary to convert none to nil. Then all the checks for
allergies? will work. If it was set to none, then allergies? would be true
which we don't want.