Add country_select gem to handle selecting a county when editing a booking.
This continues on from the work that added the html and javascript to
allow new booking to select a country of origin.

That work didn't actually have the html hooked up to the form.
This connects the html to the form so when a booking is now updated
the country code is saved on the booking.