Rename medical_conditions and associated fields -> other_information on guests and booking models.
We want guests / new bookings to be able to add extra information (that can not be handled in the new
booking form using the current radio buttons) into the the free text form element.

Currently this is only refered to as medical conditions but it will hold any extra information that
the new booking wants to add.

It will allow the new booking to write down if they have multiple allergies or dietary requirements
and also any medical conditions or anything else they think the guide should know about.