Add destiation account and platform fee to Stripe charge.
This work finishes off the basic, required integration with the
Stripe API.

Now when a charge is created (ie: when a new booking is created),
the platform fee (currently €4) is applied and transferred to our
bank account. The remainder (minus Stripe's card handling fee) is then
paid in to the guide's organisation's bank account.

We will need to manually create each guide's connected/destination
account and associate that stripe account number with the guide's
organisation.

Also, will need to, via the Stripe dashboard, add all the relevant
personal and verification information, to allow Stripe to process
payments for each guide.