Use hosted image URLs in emails.
This will fix the broken images in emails.

Created a public bucket to host public assets like these on S3.

It looks like the background image isn't working though, so that will
need to be looked at.

Also, I think we should remove the mybookingsherpa logo from the guest
email, it's not white label enough and will confuse guests.