Make the guides sign in page the root route.
Originally, we had the the guests sign in page as the default route
and you had to add guides/sign_in when you wanted to sign in as a Guide.

But that doesn't make any sense as there is no guest log in area and we
expect to only have guides logging in for a while at least.

Now when you visit the home page you will be now on the guide sign in page.