Auto focus form elements when there's an error with that field.
Part of the work after initial user testing.

This should help with the UX when there's errors.

Follow up work can use the same helper on the booking edit page.
Also can style the erroring fields so that they are highlighted for
the user, to draw their attention to it.