Highlight form fields that have user errors.
Uses Bootstrap 4 form validation functionality.

Needed to tweak the example code to work with Stripe JS (which
also checks the form input fields for valid card values and prevents
form submission).

Also needed to tweak the basic example to work with server side
validation errors.

The code: document.querySelectorAll('input[type="submit"]')[0].disabled = false;
was necessary as the submit button was remaining disabled when the fields
were then corrected so then it was not possible to submit the form even
when all fields were now valid.

refs:
https://getbootstrap.com/docs/4.0/components/forms/#validation