Check for Stripe card element before running the Stripe JS.
Part of the work in JS world was trying to get Bootstrap
custom validation working. It was necessary to make sure there
was no errors when working out why the validation wasn't working.

Previously the JS was throwing this error:
Uncaught IntegrationError: The selector you specified (#card-element)
applies to no DOM elements that are currently on the page.
Make sure the element exists on the page before calling mount().

This fixes that.