Change 'name' label to 'full name' label.
Part of the work post user testing.

One of the users only entered their first name.

The more information we can get for the guide the better.

Tried adding a placeholder text with first and last name but it
was confusing for users, so have just added full name label to
hopefully encourage users to enter their... full name.