Show past trips in separate, My Past Trips view.
Part of the work post user testing.

Trips were being ordered in descending order of start date without
any other filters.

So trips that were in the past were (at looked like) showing up at
the top of the guides trips index view, when the trip of most interest
to the guide would be the most recent upcoming trip.

This work splits the future and past trips into two separate views.

Also updated is the seeds.db so that all trips occur this year, rather
than 2020 - which looked like past trips were at the top but actually
their were future trips in Feb 2020, for ex.

We may want to add the year to trips in this view too to prevent this
confusion in future?