Set bookings' priority value before save.
This follows up to the work that just added the priority value to
bookings table.

This work actually sets the priority value based on the current
state of the booking.

That is, the priorty value will be:

0: No action required from the guide.
1: Guest has other information.
2: Booking has incomplete data which is still required.
3: Payment required.

This will then, in follow up work, be used to order the bookings
when the guide views them. With payment required at the top, incomplete
data which is still required bookings next, then if the guest has
provided other information (dietary, allergies or other information)
and finally, at the bottom, no action required for the guide on this
booking.