Order bookings by highest priority.
This is the final piece of work involving ordering bookings
based on the most pressing action required for the guide viewing
their bookings.

Bookings are now ordered based on their priority (set in a previous
commit) in the guides/trips#index view.

Subsequent work could look at optimising the DB queries involved in
this view but this work completes the main objective of this block
of work for now.