Add create and new routes and (v basic) views and actions to the Guides::Trips controller.
It's extremely likely, after initial user testing, that the first thing a Guide
will ask for is to add a new trip.

This is the basis of that work.

Follow up work will build out the views and forms and also add edit and
update actions and views to the Guides::Trips controller.