Use hardcoded Guide from email address.
PostMark, our email provider, was not delivering emails due to the
error:

'The 'From' address you supplied is not a Sender Signature on your account.

A first attempt was to try and use the PostMark API to register every new
guides's email address with them as a new sender signature.

However, if they are using a hotmail/ gmail / free email account
it is not possible to register this as an email sender signature
(due to the ability to spoof email addresses). So it has to be an
email address associated with their own domain.

For now, this is the quickest way to fix this and allow emails to be
sent to guests.

Follow up work will make sure the wording of the email doesn't get
new guests to reply to this email - as it's now replying to:
no-reply@bookyour.place