Add (hidden) edit and update guides/trips actions and views.
Follows on from this commit: 1ff0675c4d77562ce4d9369f9c38403a34995e97

It's extremely likely, after initial user testing, that the first thing a Guide
will ask for is to add a new trip.. and once they've added a new trip... the
next thing they'd want to do is edit and update that trip.

This work lets them do that.