Remove penicillin as a one of the possible allergies.
The form was breaking in an odd way with this field being present
and it seemed odd having a list of allergies that are all types of
food and then one that was penicillin.

Futher work will break this simple form element out to allow the
selection of multiple allergies without having loads of checkboxes.