Refactor External::StripeApi.
Part of the work to allow deposits to be paid.

We will need to add further integration with Stripe API to create
customers so that we can make multiple charges against a user.

That is, both deposits and the full amount.

Breaking out the creation of charges with Stripe API, to a
External::StripeApi::Charge class, means we can then add a
External::StripeApi::Customer class.

Adding the External::StripeApi::Customer class will follow this work.