Use Stripe Customer ID when making a payment, not just the one time token.
Part of the work to allow us to take deposit and then full payments from
new guests.

Now we are creating a Stripe Customer and using their id to make payment,
we can build on this work and initially take a deposit payment and then,
when the time is right collect the remainder of the full payment.