Add guides/trips/bookings#index view.
Part of the work post interview with Arno.

He said he needs to see all the bookings for a trip in one glance/view.
As a kind of spreadsheet view.

This is just the basic implementation of that view.

For now it is just a literal copy of the guides/bookings#show view.
But follow up work will build this out to be a more like a spreadsheet
style view.