Add deposit_percentage and full_payment_window_weeks to trips and organisations.
The deposit_percentage field is the percentance of a trip's full amount required
as a deposit payment.

The full_payment_window_weeks is the time in weeks that a new booking must pay the
full amount in.

So, for ex, if the full_payment_window_weeks is 6 (ie: 6 weeks) and the trip begins
in 10 weeks, then the new guest can just pay the deposit.

If the full_payment_window_weeks is 6 (ie: 6 weeks) and the trip begins
in 5 weeks, then the new guest will have to pay the full amount.

The values of these two fields on the organisation will provide default values
for new trips, where these defaults can be overriden on a trip by trip basis.