Expose deposit_percentage and full_payment_window_weeks fields in view/controller layer.
This allows guides to set the percentage deposit and window that full payment is required
for trips.

Follow up work will use this data to charge new guests either the deposit or full amount
depending upon when they are making a booking and what values the guide has set for these
fields.