Allow deposits to be paid if booking made before trip's full_payment_window.
After initial user testing with Arno, he said the ability to take deposit
payments for new bookings was a must. So needed to build this out.

This work allows the automatic collection of an initial booking payment
if the time and trip details allow it.

Further work will create a job to make the second payment for the full
amount (minus the deposit) when the full payment window had elapsed.