Rename bookings status -> payment status. Completely separate bookings' status from payment_status.
Part of the work to allow deposits to be paid separately from a single full payment.

We need to separate out all the logic relating to a booking's payment status and
its general status (allergies, medicial etc).

This work does that...

Follow up work will include a job that's run to collect the full payment of a booking
when the window for the full payment for that trip closes.